Create interface FioApi\PaymentBuilder
rename PaymentBuilder class to DomesticPaymentBuilder implementing
PaymentBuilder interface
